misc: Use more size_t when iterating over ConfArrays

+++ software/src/modules/evse/evse.cpp
@@ -606,23 +606,23 @@ void EVSE::update_all_data()
     evse_common.low_level_state.get("led_state")->updateUint(led_state);
     evse_common.low_level_state.get("cp_pwm_duty_cycle")->updateUint(cp_pwm_duty_cycle);
 
-    for (int i = 0; i < ARRAY_SIZE(adc_values); ++i)
+    for (size_t i = 0; i < ARRAY_SIZE(adc_values); ++i)
         evse_common.low_level_state.get("adc_values")->get(i)->updateUint(adc_values[i]);
 
-    for (int i = 0; i < ARRAY_SIZE(voltages); ++i)
+    for (size_t i = 0; i < ARRAY_SIZE(voltages); ++i)
         evse_common.low_level_state.get("voltages")->get(i)->updateInt(voltages[i]);
 
-    for (int i = 0; i < ARRAY_SIZE(resistances); ++i)
+    for (size_t i = 0; i < ARRAY_SIZE(resistances); ++i)
         evse_common.low_level_state.get("resistances")->get(i)->updateUint(resistances[i]);
 
-    for (int i = 0; i < ARRAY_SIZE(gpio); ++i)
+    for (size_t i = 0; i < ARRAY_SIZE(gpio); ++i)
         evse_common.low_level_state.get("gpio")->get(i)->updateBool(gpio[i]);
 
     evse_common.low_level_state.get("charging_time")->updateUint(charging_time);
     evse_common.low_level_state.get("time_since_state_change")->updateUint(time_since_state_change);
     evse_common.low_level_state.get("uptime")->updateUint(uptime);
 
-    for (int i = 0; i < CHARGING_SLOT_COUNT; ++i) {
+    for (size_t i = 0; i < CHARGING_SLOT_COUNT; ++i) {
         evse_common.slots.get(i)->get("max_current")->updateUint(max_current[i]);
         evse_common.slots.get(i)->get("active")->updateBool(SLOT_ACTIVE(active_and_clear_on_disconnect[i]));
         evse_common.slots.get(i)->get("clear_on_disconnect")->updateBool(SLOT_CLEAR_ON_DISCONNECT(active_and_clear_on_disconnect[i]));
@@ -676,7 +676,7 @@ void EVSE::update_all_data()
     user_calibration.get("voltage_div")->updateInt(voltage_div);
     user_calibration.get("resistance_2700")->updateInt(resistance_2700);
 
-    for (int i = 0; i < ARRAY_SIZE(resistance_880); ++i)
+    for (size_t i = 0; i < ARRAY_SIZE(resistance_880); ++i)
         user_calibration.get("resistance_880")->get(i)->updateInt(resistance_880[i]);
 
 #if MODULE_WATCHDOG_AVAILABLE()
